requirements.txt
altair==5.5.0
chromadb==0.6.3
langchain==0.3.23
langchain_chroma==0.2.3
langchain_openai==0.3.13
numpy==2.2.4
openai==1.75.0
pandas==2.2.3
Pillow==11.2.1
pykale==0.1.2
python-dotenv==1.1.0
pytorch_lightning==2.5.1
scikit_learn==1.6.1
streamlit==1.44.1
streamlit-card==1.0.2
streamlit-drawable-canvas==0.9.3
torch==2.6.0+cu126
torchvision==0.21.0+cu126
transformers==4.51.2